requirements.txt
Appium-Python-Client
selenium
colorama